Removed non-generic stackwrapper implementations
closes #70
