+++ StackInjector/Core/Cloning/ClonedCore.cs
@@ -13,16 +13,6 @@ namespace StackInjector.Core.Cloning
                 this.clonedCore = clonedCore;
 
 
-        public IAsyncStackWrapper ToAsyncWrapper<T> () where T : IAsyncStackEntryPoint
-        {
-            var wrapper = new AsyncStackWrapper( this.clonedCore );
-
-            this.clonedCore.entryPoint = typeof(T);
-            this.clonedCore.ServeAll();
-
-            return wrapper;
-        }
-
         public IAsyncStackWrapper<TEntry, TIn, TOut> ToAsyncWrapper<TEntry, TIn, TOut> ( AsyncStackDigest<TEntry, TIn, TOut> digest )
         {
             var wrapper = new AsyncStackWrapper<TEntry,TIn,TOut>( this.clonedCore )
@@ -36,19 +26,9 @@ namespace StackInjector.Core.Cloning
             return wrapper;
         }
 
-        public IStackWrapper<T> ToGenericWrapper<T> ()
+        public IStackWrapper<T> ToWrapper<T> ()
         {
             var wrapper = new StackWrapper<T>(this.clonedCore);
-
-            this.clonedCore.entryPoint = typeof(T);
-            this.clonedCore.ServeAll();
-
-            return wrapper;
-        }
-
-        public IStackWrapper ToWrapper<T> () where T : IStackEntryPoint
-        {
-            var wrapper = new StackWrapper( this.clonedCore );
 
             this.clonedCore.entryPoint = typeof(T);
             this.clonedCore.ServeAll();
