+++ StackInjector/Core/Cloning/IClonedCore.cs
@@ -10,25 +10,11 @@ namespace StackInjector.Core.Cloning
     {
 
         /// <summary>
-        /// convert this to an <see cref="IStackWrapper"/>
-        /// </summary>
-        /// <typeparam name="T">entry point of the new wrapper</typeparam>
-        /// <returns>the new wrapper</returns>
-        IStackWrapper ToWrapper<T> () where T : IStackEntryPoint;
-
-        /// <summary>
         /// converts this to an <see cref="IStackWrapper{TEntry}"/>
         /// </summary>
         /// <typeparam name="T">entry point of the new wrapper</typeparam>
         /// <returns>the new wrapper</returns>
-        IStackWrapper<T> ToGenericWrapper<T> ();
-
-        /// <summary>
-        /// convert this to an <see cref="IAsyncStackWrapper"/>
-        /// </summary>
-        /// <typeparam name="T">entry point of the new wrapper</typeparam>
-        /// <returns>the new wrapper</returns>
-        IAsyncStackWrapper ToAsyncWrapper<T> () where T : IAsyncStackEntryPoint;
+        IStackWrapper<T> ToWrapper<T> ();
 
 
         /// <summary>
