+++ tests/StackInjector.TEST.Async/Services/PowElaborators.cs
@@ -17,7 +17,7 @@ namespace StackInjector.TEST.Async.Services
 
 
 
-    class PowElaborator : IAsyncStackEntryPoint
+    class PowElaborator
     {
 
         [Served]
