+++ tests/StackInjector.TEST.ComplexStack/TestProgram.cs
@@ -25,7 +25,7 @@ namespace StackInjector.TEST.ComplexStack
         [Test]
         public void GenericWrapper()
         {
-            using var wrapper = Injector.OutOf<IBaseService>();
+            using var wrapper = Injector.From<IBaseService>();
 
             wrapper.Start( app => app.EntryPoint() );
         }
